Improved formatting for arrays and slices
