--- packages/react/src/SliceInput/SliceInput.utils.ts
@@ -1,10 +1,0 @@
-import { InternalTypeSchema, SliceDefinition } from '@medplum/core';
-
-export type SupportedSliceDefinition = SliceDefinition & {
-  type: NonNullable<SliceDefinition['type']>;
-  typeSchema?: InternalTypeSchema;
-};
-
-export function isSupportedSliceDefinition(slice: SliceDefinition): slice is SupportedSliceDefinition {
-  return slice.type !== undefined && slice.type.length > 0;
-}
